refactor: improve code quality and fix state cleanup issues
This commit includes several improvements to code quality, safety, and maintainability:

- Add const qualifiers to member variables that are only set during initialization
- Use std::optional for member variables that may not have a value after construction
- Fix state cleanup flags to be set immediately after cleanup operations execute
- Add Create() factory method to CompanionDeviceAuthAllInOneExecutor
- Change from shared_from_this() to weak_from_this() for better performance
- Update encode functions to return void instead of bool
- Improve subscription management with std::optional
- Add ENSURE_OR_CONTINUE null checks for safer iteration
- Remove duplicate callback protection mechanisms
- Fix uninitialized variable bug in OutboundRequest::Cancel()

These changes improve type safety, prevent accidental modification of immutable data,
and make the code more robust by ensuring resources are properly cleaned up regardless
of operation success or failure.

+++ frameworks/js/napi/src/companion_device_auth_napi_impl.cpp
@@ -61,7 +61,7 @@ napi_value CompanionDeviceAuthNapiImpl::RegisterDeviceSelectCallback(napi_env en
     deviceSelectCallback->SetCallback(callbackRef);
     int32_t ret = CompanionDeviceAuthClient::GetInstance().RegisterDeviceSelectCallback(deviceSelectCallback);
     if (ret != SUCCESS) {
-        IAM_LOGE("RegisterDeviceSelectCallback fail, ret:%{public}d", static_cast<int32_t>(ret));
+        IAM_LOGE("RegisterDeviceSelectCallback fail, ret:%{public}d", ret);
         errorCode = ret;
         return nullptr;
     }
@@ -76,7 +76,7 @@ napi_value CompanionDeviceAuthNapiImpl::UnregisterDeviceSelectCallback(napi_env 
     IAM_LOGI("start");
     int32_t ret = CompanionDeviceAuthClient::GetInstance().UnregisterDeviceSelectCallback();
     if (ret != SUCCESS) {
-        IAM_LOGE("UnregisterDeviceSelectCallback fail, ret:%{public}d", static_cast<int32_t>(ret));
+        IAM_LOGE("UnregisterDeviceSelectCallback fail, ret:%{public}d", ret);
         napi_throw(env, CompanionDeviceAuth::CompanionDeviceAuthNapiHelper::GenerateBusinessError(env, ret));
         return nullptr;
     }
@@ -124,7 +124,7 @@ napi_value CompanionDeviceAuthNapiImpl::UpdateEnabledBusinessIds(napi_env env, n
     int32_t ret =
         CompanionDeviceAuthClient::GetInstance().UpdateTemplateEnabledBusinessIds(templateId, enabledBusinessIds);
     if (ret != SUCCESS) {
-        IAM_LOGE("UpdateTemplateEnabledBusinessIds fail, ret:%{public}d", static_cast<int32_t>(ret));
+        IAM_LOGE("UpdateTemplateEnabledBusinessIds fail, ret:%{public}d", ret);
         errorCode = ret;
         return voidPromise;
     }
